requirements.txt
customtkinter
ultralytics
pillow
numpy
opencv-python
tk
